requirements.txt
clipboard
stdiomask
pycryptodome